+++ azimuth/modules/pipeline_comparison/prediction_comparison.py
@@ -1,39 +1,37 @@
-from typing import List, cast
+from typing import List
 
 from azimuth.config import ModelContractConfig
 from azimuth.dataset_split_manager import DatasetSplitManager, PredictionTableKey
 from azimuth.modules.base_classes import DatasetResultModule
 from azimuth.modules.model_contract_task_mapping import model_contract_task_mapping
 from azimuth.modules.task_execution import get_task_result
-from azimuth.types import ModuleOptions, ModuleResponse, SupportedMethod
+from azimuth.types import ModuleOptions, SupportedMethod
 from azimuth.types.pipeline_comparison import PredictionComparisonResponse
 from azimuth.types.tag import SmartTag
 from azimuth.types.task import PredictionResponse
+from azimuth.utils.validation import assert_not_none
 
 
 class PredictionComparisonModule(DatasetResultModule[ModelContractConfig]):
     def compute_on_dataset_split(self) -> List[PredictionComparisonResponse]:  # type: ignore
-        ds = self.get_dataset_split()
-        if self.config.pipelines is None or len(self.config.pipelines) < 2:
-            return [
-                PredictionComparisonResponse(one_model_disagrees=False, all_models_wrong=False)
-            ] * ds.num_rows
-        num_pipelines = len(self.config.pipelines)
+        self._validate_config()
+        num_pipelines = len(assert_not_none(self.config.pipelines))
         preds_per_pipeline = [
             self._get_predictions(pipeline_idx=pipeline_idx)
             for pipeline_idx in range(num_pipelines)
         ]
         result = []
         for predictions in zip(*preds_per_pipeline):
-            all_models_wrong = all(
+            incorrect_for_all_pipelines = all(
                 pred.postprocessed_output.preds.item() != pred.label for pred in predictions
             )
-            one_model_disagrees = (
+            pipeline_disagreement = (
                 len(set(pred.postprocessed_output.preds.item() for pred in predictions)) > 1
             )
             result.append(
                 PredictionComparisonResponse(
-                    one_model_disagrees=one_model_disagrees, all_models_wrong=all_models_wrong
+                    pipeline_disagreement=pipeline_disagreement,
+                    incorrect_for_all_pipelines=incorrect_for_all_pipelines,
                 )
             )
 
@@ -63,20 +61,27 @@ class PredictionComparisonModule(DatasetResultModule[ModelContractConfig]):
         )
         return table_key
 
-    def _save_result(self, res: List[ModuleResponse], dm: DatasetSplitManager):
-        results = cast(List[PredictionComparisonResponse], res)
-        if self.config.pipelines is None:
-            return
+    def _save_result(  # type: ignore
+        self, res: List[PredictionComparisonResponse], dm: DatasetSplitManager
+    ):
+        self._validate_config()
 
-        for pipeline_idx in range(len(self.config.pipelines)):
+        for pipeline_idx in range(len(assert_not_none(self.config.pipelines))):
             table_key = self._get_table_key_for_pipeline_index(pipeline_index=pipeline_idx)
             dm.add_tags(
                 {
                     i: {
-                        SmartTag.one_model_disagrees: response.one_model_disagrees,
-                        SmartTag.all_models_wrong: response.all_models_wrong,
+                        SmartTag.pipeline_disagreement: response.pipeline_disagreement,
+                        SmartTag.incorrect_for_all_pipelines: response.incorrect_for_all_pipelines,
                     }
-                    for i, response in enumerate(results)
+                    for i, response in enumerate(res)
                 },
                 table_key=table_key,
+            )
+
+    def _validate_config(self):
+        if self.config.pipelines is None or len(self.config.pipelines) < 2:
+            raise ValueError(
+                "Can't compute prediction comparison with less than two pipelines got"
+                f" {0 if self.config.pipelines is None else len(self.config.pipelines)}"
             )